fix: use TabViewImpl to share code across architectures
